// src/shared/enums/product-api.ts
export enum ProductAPI {
  ALL_MENS_CLOTHING = "https://fakestoreapi.com/products/category/men's%20clothing",
  ALL_WOMENS_CLOTHING = "https://fakestoreapi.com/products/category/women's%20clothing",
  ALL_ELECTRONICS = 'https://fakestoreapi.com/products/category/electronics',
  ALL_JEWELRY = 'https://fakestoreapi.com/products/category/jewelery',

  GET_PRODUCT_BY_ID = 'https://fakestoreapi.com/products/',
}